fix(changelog): keep the folding logic out of the shebang'd entry
The Windows leg of `validate` could not even parse the new test file:
`SyntaxError: Invalid or unexpected token`, no location, suite reported as
0 tests. It passed on macOS and Linux, and passed locally with CRLF line
endings, so it was neither the line endings nor the em dashes the file
contains.

The test imported `scripts/collect-changelog.mjs`, which opens with
`#!/usr/bin/env node`. Every other script test in this repo imports from
`scripts/lib/*.mjs` instead, and those files carry no shebang — the split
was already the convention here and this change did not follow it.

So the logic moves to `scripts/lib/changelog-fragments.mjs` with no
shebang, and `scripts/collect-changelog.mjs` becomes the command line
around it. The entry no longer needs its import-vs-run guard either: it is
only ever run, so it just calls main().

Stated honestly: this cannot be reproduced on macOS, so the shebang is the
best-supported explanation rather than a proven one. It matches the error,
and it is the one structural difference between this test and the ones
that pass. If Windows still fails, the next suspect is how vitest resolves
`.mjs` outside `src/`.
